refactor: MenuGroup과 Menu 패키지 분리
