+++ src/main/java/kitchenpos/menugroup/ui/MenuGroupRestController.java
@@ -1,9 +1,9 @@
-package kitchenpos.menu.ui;
+package kitchenpos.menugroup.ui;
 
 import java.net.URI;
 import java.util.List;
-import kitchenpos.menu.application.MenuGroupService;
-import kitchenpos.menu.domain.MenuGroup;
+import kitchenpos.menugroup.application.MenuGroupService;
+import kitchenpos.menugroup.domain.MenuGroup;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
