+++ src/test/java/kitchenpos/menu/application/MenuTestFixture.java
@@ -4,8 +4,8 @@ import java.math.BigDecimal;
 import java.util.List;
 import javax.transaction.Transactional;
 import kitchenpos.menu.domain.Menu;
-import kitchenpos.menu.domain.MenuGroup;
-import kitchenpos.menu.domain.MenuGroupRepository;
+import kitchenpos.menugroup.domain.MenuGroup;
+import kitchenpos.menugroup.domain.MenuGroupRepository;
 import kitchenpos.menu.domain.MenuProduct;
 import kitchenpos.menu.domain.MenuRepository;
 import kitchenpos.product.domain.ProductRepository;
