+++ src/test/java/kitchenpos/menugroup/domain/MenuGroupTest.java
@@ -1,8 +1,9 @@
-package kitchenpos.menu.domain;
+package kitchenpos.menugroup.domain;
 
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
 
 import java.util.stream.Stream;
+import kitchenpos.menugroup.domain.MenuGroup;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
