+++ src/test/java/kitchenpos/order/application/OrderServiceTest.java
@@ -11,7 +11,7 @@ import java.util.stream.Stream;
 import kitchenpos.ServiceTest;
 import kitchenpos.menu.application.MenuTestFixture;
 import kitchenpos.menu.domain.Menu;
-import kitchenpos.menu.domain.MenuGroup;
+import kitchenpos.menugroup.domain.MenuGroup;
 import kitchenpos.menu.domain.MenuProduct;
 import kitchenpos.order.domain.OrderLineItem;
 import kitchenpos.order.domain.OrderMenu;
