+++ src/test/java/kitchenpos/util/dto/SaveMenuDto.java
@@ -1,7 +1,7 @@
 package kitchenpos.util.dto;
 
 import java.util.List;
-import kitchenpos.menu.domain.MenuGroup;
+import kitchenpos.menugroup.domain.MenuGroup;
 import kitchenpos.menu.domain.MenuProduct;
 
 public class SaveMenuDto {
